requirements.txt
opencv-python>=4.8.0
mediapipe>=0.10.0
SpeechRecognition>=3.10.0
pyttsx3>=2.90
PyQt5>=5.15.0
langchain>=0.0.300
beautifulsoup4>=4.12.0
selenium>=4.12.0
pygetwindow>=0.0.9
Pillow>=10.0.0
openai-whisper>=20230918
torch>=2.0.0
transformers>=4.33.0
python-dotenv>=1.0.0
pyinstaller>=5.13.0
sounddevice>=0.4.6
numpy>=1.24.0
psutil>=5.9.0
python-whois>=0.8.0
dnspython>=2.4.0
requests>=2.31.0
